refactor(front): changed ComponentNode from several componets to one dinamic component to facilitate decision submit - add dinamic txt and memo to change component data when in preGame
